Moved pug.js to shared/
