+++ webapplication/lib/express.js
@@ -7,7 +7,7 @@ var cookieParser = require('cookie-parser');
 var path = require('path');
 var cors = require('cors');
 var config = require('../../shared/config');
-var pug = require('./pug')(config);
+var pug = require('../../shared/pug')(config);
 
 /**
  * Express configuration
