+++ webapplication/lib/layouts.js
@@ -1,10 +1,9 @@
 'use strict';
 const assert = require('assert');
-const fs = require('fs');
 const path = require('path');
 const _ = require('lodash');
 const config = require('../../shared/config');
-const pug = require('./pug')(config);
+const pug = require('../../shared/pug')(config);
 
 const helpers = require('./helpers.js');
